src/requirements.txt
Flask>1
slack_bolt
google-cloud-firestore
google-cloud-language